Rename and namespace the api test helpers.
1. ApiStubHelper --> ApiTestHelper. This rename was necessary because
future commits will introduce methods which do things apart from
stubbing.

2. Namespace the methods and include them under ActiveSupport::TestCase
so that they are available in all unit tests as instance methods.
